use solveLimbIK in ABC and fix argument type
git-svn-id: https://hrpsys-base.googlecode.com/svn/trunk@716 a991ac11-fb38-5095-8c12-a1ddb0715245

+++ rtc/ImpedanceController/JointPathEx.cpp
@@ -324,7 +324,7 @@ void JointPathEx::solveLimbIK (const hrp::Vector3& _vel_p,
                                const int transition_count,
                                const double avoid_gain,
                                const double reference_gain,
-                               const int MAX_TRANSITION_COUNT,
+                               const double MAX_TRANSITION_COUNT,
                                const hrp::dvector& qrefv,
                                bool DEBUGP)
 {
