+++ rtc/ImpedanceController/JointPathEx.h
@@ -27,7 +27,7 @@ namespace hrp {
                       const int transition_count,
                       const double avoid_gain,
                       const double reference_gain,
-                      const int MAX_TRANSITION_COUNT,
+                      const double MAX_TRANSITION_COUNT,
                       const hrp::dvector& qrefv,
                       bool DEBUGP = false);
   protected:
